Imported systemc_tests-2.2.05jun06_beta
Summary: Imported systemc_tests-2.2.05jun06_beta
Keywords:

Imported systemc_tests-2.2.05jun06_beta
into Git repository

+++ tests/systemc/tracing/wif_trace/mixed/isaac.h
@@ -32,7 +32,7 @@
 */
 
 
-typedef unsigned long int UINT32;
+typedef unsigned int UINT32;
 const UINT32 GOLDEN_RATIO = UINT32(0x9e3779b9);
